serverfastapi/requirements.txt
fastapi==0.103.1
python-socketio==5.9.0
uvicorn==0.23.2
websockets==11.0.3
pymongo==3.11.0
pytz
dnspython 